// src/app/interfaces/item-pedido-compra.ts
export interface ItemPedidoCompra {
    idEmpresa: number;
    idPedidoCompra: number;
    idItemPedidoCompra: number;
    idProduto: number;
    quantidade: number;
    precoUnitario: number;
    total: number;
    codigo: string;
    descricao: string;
    tipoProduto: string;
    unidadeMedida: string;
}